BF-helper is a brainfuck utility written in Python
The idea is to help you see the state of the brainfuck array when you
are coding.

Currently there is a prompt where you can put a string that will be
evaluated as brainfuck code, or you can type "p" to see the values of
the array.

To avoid having to print a huge number of values on the screen, the
print command only prints the cells where the pointer has passed at
least once. So don't be afraid if the print command only shows 10
cells of the array.

Type "q" to quit.